Move CanonicalPageInfo query into internal namespace
Making it more clear that this is a query for internal usage
