+++ modules/wikis/app/services/wikis/adapters/providers/xwiki/queries/internal/canonical_page_info.rb
@@ -33,41 +33,41 @@ module Wikis
     module Providers
       module XWiki
         module Queries
-          class SearchPages < BaseQuery
-            include Concerns::XWikiQuery
+          module Internal
+            # Fetch page information using a canonical XWiki identifier
+            class CanonicalPageInfo < BaseQuery
+              include Concerns::XWikiQuery
 
-            # Limiting result size rather strictly, because each result will cause another HTTP call to XWiki, this does not
-            # scale well. A stricter limit improves the worst case latency.
-            MAXIMUM_RESULTS = 20
+              def call(input_data:, auth_strategy:)
+                ref = CanonicalPageReference.parse(input_data.identifier)
+                return failure(code: :not_found) unless ref
 
-            def call(input_data:, auth_strategy:)
-              query = { q: "\"#{escape_quotes input_data.query}\"", number: MAXIMUM_RESULTS }
-
-              authenticated(auth_strategy) do |http|
-                handle_response(http.get(rest_url("wikis/query", query:))) do |json|
+                perform_request(ref, auth_strategy:) do |data|
                   success(
-                    fetch_json(json, "searchResults")
-                      .uniq { |r| fetch_json(r, "id") }
-                      .map do |r|
-                        result = canonical_page_info(identifier: fetch_json(r, "id"), auth_strategy:)
-                        return result if result.failure?
-
-                        result.value!
-                      end
+                    Results::PageInfo.new(
+                      identifier: StablePageReference.parse(fetch_json(data, "id")).to_s,
+                      title: data.fetch("title"),
+                      href: data.fetch("xwikiAbsoluteUrl"),
+                      provider:
+                    )
                   )
                 end
               end
-            end
 
-            private
-
-            def escape_quotes(string)
-              string.gsub("\\", "\\\\").gsub('"', '\"')
-            end
-
-            def canonical_page_info(identifier:, auth_strategy:)
-              Input::PageInfo.build(identifier:).bind do |input_data|
-                Internal::CanonicalPageInfo.new(model: provider).call(input_data:, auth_strategy:)
+              def perform_request(reference, auth_strategy:, &)
+                authenticated(auth_strategy) do |http|
+                  handle_response(
+                    # This query is implemented as a PUT on the XWiki side, because it dynamically creates the
+                    # stable identifier that it returns. Passing an empty JSON body is also required for the
+                    # endpoint to not raise an error 🤷
+                    http.with(headers: { "Content-Type": "application/json" })
+                        .put(
+                          rest_url("openproject/documents", query: { docRef: reference.to_s }),
+                          body: "{}"
+                        ),
+                    &
+                  )
+                end
               end
             end
           end
